+++ HISTORY.rst
@@ -124,7 +124,9 @@
 * Add document state action view test.
 * Remove sidebar menu instance. The secondary menu and the
   previour sidebar menu now perform the same function.
-
+* Backport source column identifiable and sortable 
+  improvements.
+  
 3.1.11 (2019-04-XX)
 ===================
 * Fix multiple tag selection wizard step.